feat(return_label): add success? method to ReturnLabel class
This commit introduces a new method `success?` in the `ReturnLabel` class to provide a more idiomatic way to check the success status of a return label. Additionally, the `reset` method was removed from the `AnPostReturn` module as it was deemed unnecessary. The corresponding test cases were updated to include assertions for the new `success?` method.